M5b: Ticker Tape and Price Chart; Floor shows last trade and today's average
Price history now keeps quarter-day bars with traded value (v2; v1 loads), so a
chart shows a price reacting within the day. PriceChart: 7 days x 4 points with
high-low per quarter, gaps carried forward, a sparkline of daily closes, packed
into item data. Ticker Tape block (node $1,500 after the Trading Floor; blueprint
1 Clockwork Gear, 4 iron, 1 ink) prints a chart of any book for 1 Ledger Paper +
1 Ink Bottle; right-click the chart for a line chart. Guide Reading a Chart, quest
Read the Tape. Floor screen (236 px): 'Last trade' plus today's volume-weighted
average and low-high. 166 unit tests, 46 GameTests.

+++ sim/src/main/java/com/realisticmarkets/sim/TraderSim.java
@@ -77,7 +77,7 @@ public final class TraderSim {
             for (int i = 0; i < n; i++) {
                 double before = day;
                 day += DAY_STEP;
-                for (FloorCatalog.Book b : floor.catalog().all()) floor.auction(b.item(), fair(b.item()), DAY_STEP, (long) day);
+                for (FloorCatalog.Book b : floor.catalog().all()) floor.auction(b.item(), fair(b.item()), DAY_STEP, day);
                 if (Math.floor(day + 1e-9) > Math.floor(before + 1e-9)) floor.dawn((long) Math.floor(day + 1e-9));
             }
         }
